pyproject.toml
[project]
name = "zmq-docker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pre-commit>=4.1.0",
    "pyzmq>=26.3.0",
]
